Extracted styles to separate files
